# backend/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kashtbhanjan Medical Shop Management System API",
    "description": "Enterprise-grade Pharmacy ERP API documentation covering Sales POS, FEFO Inventory, Suppliers, Customers, and System Maintenance Utilities.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token to authorize requests."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Invalid request parameters" }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate User",
        "description": "Returns a JWT token for authenticated admin or staff.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "admin@kashtbhanjan.com" },
                  "password": { "type": "string", "example": "admin123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsIn..." },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "example": "664c12ef0d27..." },
                        "name": { "type": "string", "example": "Krupesh Admin" },
                        "role": { "type": "string", "example": "admin" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid email or password",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}