# backend/controllers/authController.js
const jwt = require('jsonwebtoken');
const User = require('../models/User');
const ActiveSession = require('../models/ActiveSession');
const LoginHistory = require('../models/LoginHistory');
const {
  validatePasswordStrength,
  isAccountLocked,
  handleFailedLogin,
  handleSuccessfulLogin,
  forceLogoutAllUserSessions
} = require('../config/SecurityService');
const { logSystemAction } = require('../config/AuditService');

// Helper to generate JWT Token
const generateToken = (id) => {
  return jwt.sign({ id }, process.env.JWT_SECRET, {
    expiresIn: process.env.JWT_EXPIRE || '24h'
  });
};

// @desc    Auth user & get token
// @route   POST /api/auth/login
// @access  Public
const loginUser = async (req, res, next) => {
  const { email, password } = req.body;
  const ipAddress = req.ip || req.headers['x-forwarded-for'] || req.connection.remoteAddress || '127.0.0.1';
  const userAgent = req.headers['user-agent'] || 'Unknown';

  try {
    const user = await User.findOne({ email });

    if (!user) {
      await handleFailedLogin(email, ipAddress, userAgent);
      return res.status(401).json({ success: false, message: 'Invalid credentials' });
    }

    // Check if account is locked
    const locked = await isAccountLocked(user);
    if (locked) {
      return res.status(423).json({
        success: false,
        message: 'Account is locked due to multiple failed attempts. Please try again after 15 minutes.'
      });
    }

    // Check password match
    const isMatch = await user.matchPassword(password);
    if (!isMatch) {
      const status = await handleFailedLogin(email, ipAddress, userAgent);
      if (status === 'Locked') {
        return res.status(423).json({
          success: false,
          message: 'Account has been locked for 15 minutes due to 5 consecutive failed login attempts.'
        });
      }
      return res.status(401).json({ success: false, message: 'Invalid credentials' });
    }

    // Generate token
    const token = generateToken(user._id);

    // Register active session & concurrent limits
    const session = await handleSuccessfulLogin(user, ipAddress, userAgent, token);

    // Log action
    await logSystemAction({ user, ip: ipAddress, headers: { 'user-agent': userAgent }, method: 'POST', originalUrl: '/api/auth/login' }, {
      actionType: 'User Login',
      module: 'Security',
      entityType: 'User',
      entityId: user._id,
      remarks: `Successful login from browser ${session.deviceInfo.browser}`
    });

    res.json({
      success: true,
      token,
      user: {
        id: user._id,
        name: user.name,
        email: user.email,
        role: user.role
      }
    });
  } catch (error) {
    next(error);
  }
};

// @desc    Logout active session
// @route   POST /api/auth/logout
// @access  Private
const logoutUser = async (req, res, next) => {
  try {
    const token = req.headers.authorization && req.headers.authorization.split(' ')[1];
    
    if (token) {
      const session = await ActiveSession.findOne({ sessionToken: token });
      if (session) {
        session.isRevoked = true;
        await session.save();
        
        await logSystemAction(req, {
          actionType: 'User Logout',
          module: 'Security',
          entityType: 'ActiveSession',
          entityId: session._id,
          remarks: `User session logged out successfully.`
        });
      }
    }

    res.json({ success: true, message: 'Logged out successfully' });
  } catch (error) {
    next(error);
  }
};

// @desc    Get user profile
// @route   GET /api/auth/profile
// @access  Private
const getUserProfile = async (req, res, next) => {
  try {
    const user = await User.findById(req.user.id);
    if (user) {
      res.json({
        success: true,
        user: {
          id: user._id,
          name: user.name,
          email: user.email,
          role: user.role,
          createdAt: user.createdAt
        }
      });
    } else {
      res.status(404).json({ success: false, message: 'User not found' });
    }
  } catch (error) {
    next(error);
  }
};

// @desc    Get all active sessions for current user
// @route   GET /api/auth/sessions
// @access  Private
const getActiveSessions = async (req, res, next) => {
  try {
    const sessions = await ActiveSession.find({ userId: req.user.id, isRevoked: false })
      .sort({ loginTime: -1 })
      .lean();
    res.json({ success: true, sessions });
  } catch (error) {
    next(error);
  }
};

// @desc    Force revoke single session by ID
// @route   POST /api/auth/sessions/revoke
// @access  Private
const revokeSession = async (req, res, next) => {
  try {
    const { sessionId } = req.body;
    const session = await ActiveSession.findOne({ _id: sessionId, userId: req.user.id });
    
    if (!session) {
      return res.status(404).json({ success: false, message: 'Active session not found.' });
    }

    session.isRevoked = true;
    await session.save();

    await logSystemAction(req, {
      actionType: 'User Session Revoked',
      module: 'Security',
      entityType: 'ActiveSession',
      entityId: session._id,
      remarks: `User manually terminated active session from terminal.`
    });

    res.json({ success: true, message: 'Session revoked successfully.' });
  } catch (error) {
    next(error);
  }
};

// @desc    Force revoke all user sessions
// @route   POST /api/auth/sessions/revoke-all
// @access  Private
const revokeAllSessions = async (req, res, next) => {
  try {
    await forceLogoutAllUserSessions(req.user.id, 'User manual revoke all request');
    res.json({ success: true, message: 'All active sessions terminated successfully.' });
  } catch (error) {
    next(error);
  }
};

// @desc    Reset password with strength validation and force logouts
// @route   POST /api/auth/reset-password
// @access  Private
const resetPassword = async (req, res, next) => {
  try {
    const { newPassword } = req.body;
    
    if (!newPassword) {
      return res.status(400).json({ success: false, message: 'New password is required.' });
    }

    const isValid = validatePasswordStrength(newPassword);
    if (!isValid) {
      return res.status(400).json({
        success: false,
        message: 'Password does not meet complexity rules: Minimum 8 characters, with 1 uppercase, 1 lowercase, 1 numeric, and 1 special symbol.'
      });
    }

    const user = await User.findById(req.user.id);
    user.password = newPassword; // Pre-save hook hashes this
    await user.save();

    // Log to audit log
    await logSystemAction(req, {
      actionType: 'User Password Reset',
      module: 'Security',
      entityType: 'User',
      entityId: user._id,
      remarks: 'Password reset completed. Force terminating other sessions.'
    });

    // Force terminate all active sessions for this user
    await forceLogoutAllUserSessions(user._id, 'Password reset trigger');

    res.json({ success: true, message: 'Password updated successfully. Other active sessions have been logged out.' });
  } catch (error) {
    next(error);
  }
};

// @desc    Change user role and force logout sessions (Admin only)
// @route   POST /api/auth/change-role
// @access  Private/Admin
const changeUserRole = async (req, res, next) => {
  try {
    if (req.user.role !== 'admin') {
      return res.status(403).json({ success: false, message: 'Admin role required.' });
    }

    const { targetUserId, newRole } = req.body;
    if (!targetUserId || !newRole) {
      return res.status(400).json({ success: false, message: 'Target user ID and new role are required.' });
    }

    if (!['admin', 'staff'].includes(newRole)) {
      return res.status(400).json({ success: false, message: 'Invalid role selection.' });
    }

    const user = await User.findById(targetUserId);
    if (!user) {
      return res.status(404).json({ success: false, message: 'User not found.' });
    }

    const oldRole = user.role;
    user.role = newRole;
    await user.save();

    // Log to audit
    await logSystemAction(req, {
      actionType: 'User Role Updated',
      module: 'Security',
      entityType: 'User',
      entityId: user._id,
      oldValues: { role: oldRole },
      newValues: { role: newRole },
      remarks: `Admin changed role of user ${user.email} from ${oldRole} to ${newRole}.`
    });

    // Force logout all sessions for that user
    await forceLogoutAllUserSessions(user._id, 'Role classification changed');

    res.json({ success: true, message: `User role changed successfully to ${newRole}. User sessions have been terminated.` });
  } catch (error) {
    next(error);
  }
};

module.exports = {
  loginUser,
  logoutUser,
  getUserProfile,
  getActiveSessions,
  revokeSession,
  revokeAllSessions,
  resetPassword,
  changeUserRole
};